Cleaned up the read data and alignment tools a bit.

+++ src/core/functions/builtin/Func_dppSeat.cpp
@@ -84,7 +84,7 @@ RbPtr<RbLanguageObject> Func_dppSeat::executeFunction( void ) {
 					j++;
 				}
 				tables[j] += 1;
-				clusters[i] = j-1;
+				clusters[i] = (int)(j - 1);
 			}
 		}
 	}
